bdd/pyproject.toml
[project]
name = "tests"
description = "functional module"
version = "1.0.0"
authors = [
    { name = "Thierno", email = "thierno@minkadi.com" }
]
dependencies = [
    "pydantic==2.7.4",
    "pydantic-settings",
    "loguru==0.6.0",
    "psutil==5.9.4",
    "pynvml==11.5.0",
    "behave",
    "python-keycloak",
    "sqlalchemy",
    "psycopg2-binary",
    "python-dateutil"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]
